Frozen todo check

+++ apps/server/src/todos/todos.service.ts
@@ -96,6 +96,17 @@ export class TodosService {
   }
 
   async create(ownerId: string, data: CreateTodoInput): Promise<Todo> {
+    if (data.parent?.id) {
+      const rootParent = await this.todosRepository.findTodoRootParent(
+        data.parent.id,
+      );
+      if (rootParent.frozen) {
+        throw new BadRequestException(
+          'Cannot add a new task under frozen parent',
+        );
+      }
+    }
+
     const created = await this.todosRepository.createTodo({
       title: data.title,
       ownerId: ownerId,
@@ -115,7 +126,7 @@ export class TodosService {
   async update(id: string, data: UpdateTodoInput): Promise<Todo | undefined> {
     const rootParent = await this.todosRepository.findTodoRootParent(id);
 
-    if (id === rootParent.id) {
+    if (id === rootParent?.id) {
       if (rootParent.frozen && data.frozen === undefined) {
         throw new BadRequestException('Cannot update frozen task');
       }
@@ -148,6 +159,10 @@ export class TodosService {
   }
 
   async delete(id: string): Promise<boolean> {
+    const rootParent = await this.todosRepository.findTodoRootParent(id);
+    if (rootParent?.frozen) {
+      throw new BadRequestException('Cannot delete todo under frozen task');
+    }
     const deletedTodo = await this.todosRepository.deleteTodo({ id: id });
     if (deletedTodo) {
       await this.pubSubService.publish<TodoSubscriptionMessage>(
